se agrego la segunda parte de la entrega correspondiente a la semana 5 (Main.java), generar reportes finales de ventas
