+++ Product.csv
@@ -1,0 +1,11 @@
+IDProducto;NombreProducto;PrecioPorUnidad
+5021735246;Atornillador;80457
+8960912927;Arandela;50101
+3933099155;Llave;19197
+7948695162;Tornillo;23347
+8831279410;Destornillador;22452
+9259568915;Broca;60601
+8057936992;Tuerca;17734
+1227531212;Serrucho;38391
+1797389817;Taladro;41307
+7136017708;Martillo;21120
