+++ Reporte_Ventas_Productos.csv
@@ -1,0 +1,11 @@
+Nombre_Producto;Precio_Unidad;Cantidad_Vendida
+Arandela;50101;149
+Serrucho;38391;130
+Tornillo;23347;129
+Martillo;21120;119
+Llave;19197;118
+Tuerca;17734;102
+Broca;60601;100
+Taladro;41307;99
+Destornillador;22452;91
+Atornillador;80457;86
