+++ Reporte_Ventas_Vendedores.csv
@@ -1,0 +1,11 @@
+Nombre_Vendedor;Apellido_Vendedor;Total_Ventas
+Jhon;Rivera;5228553
+Sara;Rivera;5164135
+Jhon;Contreras;4668174
+Eduar;Cardona;4492014
+Estefany;Rivera;4468647
+Estefany;Jimenes;3942793
+Michael;Alvares;3828841
+Camilo;Alvares;3398063
+Andrea;Rivera;3228258
+Jhon;Cardona;2747485
