+++ Sales_CC_1200452094.csv
@@ -1,0 +1,11 @@
+ID_Producto;Cantidad_Producto_vendido
+1227531212;12
+8960912927;7
+8057936992;15
+8057936992;3
+8057936992;18
+9259568915;16
+5021735246;14
+7136017708;15
+9259568915;1
+9259568915;9
