+++ Sales_CC_1285475978.csv
@@ -1,0 +1,11 @@
+ID_Producto;Cantidad_Producto_vendido
+9259568915;9
+1227531212;16
+9259568915;8
+1227531212;20
+5021735246;2
+8831279410;9
+8057936992;11
+5021735246;8
+8831279410;9
+7136017708;6
