+++ Sales_CC_1355312535.csv
@@ -1,0 +1,11 @@
+ID_Producto;Cantidad_Producto_vendido
+3933099155;18
+5021735246;16
+1797389817;15
+3933099155;20
+8057936992;8
+9259568915;7
+7948695162;15
+7948695162;1
+8831279410;20
+7948695162;20
