+++ Sales_CE_1232292593.csv
@@ -1,0 +1,11 @@
+ID_Producto;Cantidad_Producto_vendido
+5021735246;6
+1227531212;6
+7136017708;2
+8960912927;20
+7136017708;18
+7136017708;1
+8057936992;9
+7948695162;16
+7948695162;9
+3933099155;17
