+++ Sales_CE_1448251509.csv
@@ -1,0 +1,11 @@
+ID_Producto;Cantidad_Producto_vendido
+8831279410;15
+7136017708;10
+1227531212;2
+7136017708;1
+3933099155;10
+7948695162;19
+8831279410;20
+7136017708;6
+8960912927;20
+9259568915;16
