+++ Sales_CE_1816821320.csv
@@ -1,0 +1,11 @@
+ID_Producto;Cantidad_Producto_vendido
+8057936992;11
+1227531212;17
+1227531212;18
+9259568915;6
+1797389817;20
+7948695162;7
+8057936992;11
+5021735246;11
+7136017708;9
+8960912927;20
