+++ Sales_TI_1288810713.csv
@@ -1,0 +1,11 @@
+ID_Producto;Cantidad_Producto_vendido
+8960912927;19
+1797389817;15
+3933099155;13
+7948695162;7
+8831279410;4
+8960912927;2
+3933099155;6
+1227531212;4
+1227531212;5
+8831279410;5
