+++ Sales_TI_1652125140.csv
@@ -1,0 +1,11 @@
+ID_Producto;Cantidad_Producto_vendido
+9259568915;11
+8960912927;15
+1227531212;16
+1797389817;19
+8960912927;11
+1797389817;16
+1797389817;10
+8960912927;7
+7948695162;10
+8831279410;9
