+++ Sales_TI_1872849724.csv
@@ -1,0 +1,11 @@
+ID_Producto;Cantidad_Producto_vendido
+5021735246;2
+7136017708;16
+7948695162;15
+7948695162;10
+7136017708;4
+8057936992;16
+1227531212;5
+3933099155;14
+1227531212;1
+5021735246;18
