+++ Sales_TI_1882836874.csv
@@ -1,0 +1,11 @@
+ID_Producto;Cantidad_Producto_vendido
+8960912927;17
+9259568915;17
+7136017708;6
+1227531212;8
+1797389817;4
+3933099155;20
+7136017708;11
+7136017708;14
+5021735246;9
+8960912927;11
