+++ Salesman.csv
@@ -1,0 +1,11 @@
+TipoDocumento;NúmeroDocumento;NombresVendedor;ApellidosVendedor
+CE;1232292593;Andrea;Rivera
+CE;1816821320;Sara;Rivera
+TI;1882836874;Jhon;Contreras
+CC;1200452094;Estefany;Rivera
+CC;1285475978;Estefany;Jimenes
+CE;1448251509;Michael;Alvares
+TI;1288810713;Jhon;Cardona
+TI;1872849724;Camilo;Alvares
+TI;1652125140;Jhon;Rivera
+CC;1355312535;Eduar;Cardona
