--- products.csv
@@ -1,10 +1,0 @@
-P1;Producto1;56
-P2;Producto2;54
-P3;Producto3;93
-P4;Producto4;47
-P5;Producto5;94
-P6;Producto6;92
-P7;Producto7;93
-P8;Producto8;30
-P9;Producto9;48
-P10;Producto10;80
