--- reporteProductos.csv
@@ -1,10 +1,0 @@
-Producto9;17.0;8
-Producto8;46.0;6
-Producto6;60.0;5
-Producto2;64.0;4
-Producto4;71.0;2
-Producto10;5.0;2
-Producto1;80.0;0
-Producto3;68.0;0
-Producto5;84.0;0
-Producto7;36.0;0
