--- reporteVendedores.csv
@@ -1,5 +1,0 @@
-Maria Gomez;1120.0
-Luis Perez;0.0
-Luis Diaz;0.0
-Juan Perez;0.0
-Pedro Lopez;0.0
