--- sales_Carlos_1001.csv
@@ -1,9 +1,0 @@
-CC;1001
-P6;1
-P7;5
-P8;2
-P7;5
-P6;4
-P1;4
-P7;5
-P2;5
