--- salesmen_info.csv
@@ -1,5 +1,0 @@
-CC;1001;Pedro;Gomez
-CC;1002;Luis;Diaz
-CC;1003;Pedro;Diaz
-CC;1004;Juan;Gomez
-CC;1005;Juan;Rodriguez
